factor: start eval step
Mostly appears to work for addition and multiplication. The logic is
slightly wrong for both subtraction and division. There is a bug with
subtraction that requires further investigation.

A lot of refactoring is probably needed, especially for the approach of
tagging every value with its type. This becomes very unwieldy. Maybe
a wrapper to apply a word into a pair would be good here.